requirements.txt
pytest
pytest-xdist
requests
allure-pytest
pytest-html